+++ tests/characters/control/blockchain/test_rpc_control_blockchain.py
@@ -1,14 +1,30 @@
 import pytest
 
-from nucypher.characters.control.specifications import AliceSpecification, BobSpecification, EnricoSpecification
+from nucypher.characters.control.interfaces import AliceInterface, BobInterface, EnricoInterface
 
-alice_specification = AliceSpecification()
-bob_specification = BobSpecification()
-enrico_specification = EnricoSpecification()
+alice_specification = AliceInterface.specifications
+bob_specification = BobInterface.specifications
+enrico_specification = EnricoInterface.specifications
+
+
+def get_fields(specification, method_name):
+
+    spec = specification[method_name]
+    input_fields = [k for k, f in spec.load_fields.items() if f.required]
+    optional_fields = [k for k, f in spec.load_fields.items() if not f.required]
+    required_output_fileds = list(spec.dump_fields.keys())
+
+    return (
+        input_fields,
+        optional_fields,
+        required_output_fileds
+    )
 
 
 def validate_json_rpc_response_data(response, method_name, specification):
-    _input_fields, _optional_fields, required_output_fileds = specification.get_specifications(interface_name=method_name)
+
+    required_output_fileds = get_fields(specification, method_name)[-1]
+
     assert 'jsonrpc' in response.data
     for output_field in required_output_fileds:
         assert output_field in response.content
@@ -23,7 +39,7 @@ def test_alice_rpc_character_control_create_policy(alice_rpc_test_client, create
     assert rpc_response.success is True
     assert rpc_response.id == 1
 
-    _input_fields, _optional_fields, required_output_fileds = alice_specification.get_specifications(interface_name=method_name)
+    _input_fields, _optional_fields, required_output_fileds = get_fields(alice_specification, method_name)
 
     assert 'jsonrpc' in rpc_response.data
     for output_field in required_output_fileds:
